refactor(dungeon-builder): extract useBoardEditing, lazy-init cst state
useBoardEditing was already a cleanly self-contained hook (palette/
placement selection + handlePlace/handleMove/handleDelete/etc.) but
lived inline in DungeonBuilderConcept.tsx instead of its own module.
Moved verbatim to useBoardEditing.ts; CreationBoard.tsx/
CreationConcept.tsx now import the BoardEditing type from there
directly instead of through DungeonBuilderConcept.tsx.

Also fixed the two per-render parseDungeon(...) calls the same audit
flagged: `const initial = parseDungeon(SHOWCASE_YAML)` (and the
creation-mode equivalent) ran on EVERY render since it wasn't wrapped
in a useState lazy initializer — React only uses a useState argument
on the first render, but a bare expression still gets evaluated every
time. Wrapped both in `useState(() => parseDungeon(...))`, and did the
same for the adjacent `serializeDungeon(initial.cst)` calls feeding
the initial yamlText state, since they have the identical shape of the
same bug in the same 4-line block.

Verified live: both edit and creation mode load correctly, and placing
a prop through the extracted hook still updates the board/YAML
correctly (pillar count incremented, new placement synced).

typecheck clean; full suite 1772/1772 passing.
